Create README.md
It's a web application which has been developed using HTML , CSS and JavaScript in which a user can search Super heroes by their name and can add them as their favorite and also can know more about them like their powers and other stuffs. The index.html file makes the main page on which we search for the super hero, styling for the main page is provided by style.css file and functionality is provided by the scirpt.js file. The more-info.html file provides the informational page for the superheores to be searched, more-info.css provides the stylings to the page, and more-info.js file provides the functionality to the page. The favorites.html file stores the information of the favorite superhereos of an individual, the css and js file are also having the same name as of the html file.